It would be funny to watch me lose my sanity over time
