+++ src/types/views/index.d.ts
@@ -1,0 +1,3 @@
+export interface Message {
+    type: string;
+}
