+++ webview/actions/main.tsx
@@ -1,7 +1,7 @@
 import { render } from 'preact';
 import { useState, useEffect, useMemo, useId } from 'preact/hooks';
-import type { ActionNode, ActionsViewMessage, ActionsViewProviderMessage, SettingsContext } from '@/views/actions';
-import { PermissionLevel } from '@/config';
+import type { ActionNode, ActionsViewMessage, ActionsViewProviderMessage, SettingsContext } from '@typing/views/actions';
+import { PermissionLevel } from '../../src/types/actions';
 
 interface State {
     actions: ActionNode[];
@@ -188,7 +188,7 @@ function ActionsView() {
                 <p><span class="disabled">Grayed out = Permission is not registered</span></p>
                 <p>Some actions may have secondary registration conditions and not be registered even if permission is granted.</p>
                 <p>
-                    {levels.map(({level, icon, title, description}) =>
+                    {levels.map(({ level, icon, title, description }) =>
                         <div class="permission-level-description">
                             <div class="permission-level-radio">
                                 <label data-show-checked data-permission-level={level}>
